kg816/818: Fix query string processing
Sometime between legacy and now, the strings in wouxun.py were changed
to bytes. Most radios in this file define a tuple of bytes to be sent to
the radio....if one doesn't work, the next one is tried. For the
816/818, only one bytes object is defined. The code that processes this
checked if the input was a `str`. It failed on a `bytes` object and
started processing the input as an iterable.

Related to #11208

This is why the person in issue 11208 got an error reading their radio,
even though chirp doesn't directly support it.